requirements.txt
fastapi
google-api-python-client
lxml
pycryptodome
uvicorn